requirements.txt
Django==3.1.7
djangorestframework==3.12.4
psycopg2==2.8.6
requests==2.25.1
scipy